airflow/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-apache-spark
pyarrow
pyspark
pytest
boto3
requests
pycryptodome
pytz
moto